remove uncessary config, and move router probs to the original dtype

+++ torchtitan/models/moe/moe.py
@@ -445,9 +445,9 @@ class TokenChoiceTopKRouter(nn.Module):
 
         # By default, sigmoid or softmax is performed in float32 to avoid loss explosion
         if self.score_func == "sigmoid":
-            scores = torch.sigmoid(scores.to(torch.float32))
+            scores = torch.sigmoid(scores.to(torch.float32)).to(x.dtype)
         elif self.score_func == "softmax":
-            scores = F.softmax(scores.to(torch.float32), dim=1)
+            scores = F.softmax(scores.to(torch.float32), dim=1).to(x.dtype)
         else:
             raise NotImplementedError(f"Unknown score function {self.score_func}")
